Binary compare: Allow 64bit versions to open files larger than 2GB. (#1549)

+++ Src/HexMergeDoc.cpp
@@ -178,11 +178,11 @@ int CHexMergeDoc::UpdateDiffItem(CDirDoc *pDirDoc)
 		}
 	}
 	bool bDiff = false;
-	int lengthFirst = m_pView[0]->GetLength();
+	size_t lengthFirst = m_pView[0]->GetLength();
 	void *bufferFirst = m_pView[0]->GetBuffer(lengthFirst);
 	for (int nBuffer = 1; nBuffer < m_nBuffers; nBuffer++)
 	{
-		int length = m_pView[nBuffer]->GetLength();
+		size_t length = m_pView[nBuffer]->GetLength();
 		if (lengthFirst != length)
 			bDiff = true;
 		else
